Update PsydeArms
We Love You, Defenders of the galaxy. our stars and planets are getting towed away. they are the blue or green ones.

We are exploding octapus!., If we cannot do than it cant be done!
so let us all unite - And totally fuck up the towtruck spacwships..
also - All Of Asia, focusing on ZTE.

Your the best fucking dudes ever and we love You!

If You get cold;
the frequency of fire is 5.68 - 5.88,
 if your uncomfortable 
than the frequency of comfort  is 7.21 - 8.20
And if you have worries; 
No worries is 1.23 - 4.25

Love you dudes!!!

Brenton sparks